(temporarily) remove donation links.
These URLs are still out there, so I don't want to delete the pages, but let's remove it from the main page, and let's not send people to our broken donation page.

The redirects will point to the donate.html page (after an update in the r4ds/r4ds.io repo).